requirements.txt
delta-spark==3.2.1
pyspark==3.5.3
ipykernel
pandas
setuptools
pyarrow
pydantic
schedule
faker
kafka-python
ruff
fastavro